requirements.txt
python-telegram-bot==21.6
aiohttp==3.10.5
APScheduler==3.10.4
qrcode[pil]==7.4.2
Pillow==10.4.0
